Fix resolve_getitem_callee -> resolve_base
Summary:
In this commit I start making the inner helpers of
`parse_callable_if_appropriate` suitable for use in a world with subscripts...
this means we need to always separately match the case of a `__getitem__` call
(which will be transformed into a subscript call when we cut the parser over);
in the process we have to update type signatures anywhere where we've assumed
that function calls and subscripts have the same type structure (becasue after
the cutover they won't!).

This particular commit fixes `resolve_getitem_callee`, instead making it work
on the base inside of the attribute representign `__getitem__`; this is important
because once we cut over we will *only* have the base, no "callee" will exist.

Reviewed By: samwgoldman

Differential Revision: D57072634

fbshipit-source-id: 1e02ea1866534be4b21afd6225928505efa9346b