Fix pretty-printing of Callable types
Summary:
We had some pretty wild bare-getitem logic going on here that is
definitely not compatible with a future of proper `Subscript`.

This commit modifies both the test of that aspect of pretty-printing
and the underlying pretty-printer to properly use list syntax, instead
of using a top-level `__getitem__` *function* (which very much is not
a thing at runtime!) as a pun for a list.

Reviewed By: grievejia

Differential Revision: D57078069

fbshipit-source-id: 27189fad0620b18a45e65a02f40c9aa5a6ee7c4b