Fix is_typing_callable to work on subscript base
Summary:
I'm continuing to refactor the internals of the `parse_callable_if_appropriate`
portion of `create_logic` so that we don't have to pass in a call expression /
callee anymore, which is required in order to eliminate `__getitem__` lowering
from the parser.

This commit cuts over the second of the three inner helpers, the
`is_typing_callable` function, and tweaks the block at the end to defer
reconstructing a "resolved_callable" until `parse_callable` (which hasn't
yet been converted).

Reviewed By: samwgoldman

Differential Revision: D57072625

fbshipit-source-id: 9ac8451af9adde49f5f8c47f290e707156ec7062