Lift subscript parsing out of main match
Summary:
In an earlier diff of this stack, I consolidated all of the `__getitem__`
cases of expression-to-type conversions into one outer match statement,
using an inner match to handle various special cases and then the fallback
to either Callable or normal parametric types.

In my opinion one of the problems we run into in Pyre is too much use
of nested match statements; the problem with these is not that they
are unreadable per se but that we accumulate more and more scope as we
nest logic which makes it hard to find what the actual dependencies of
any given bit of logic are.

So, in this diff I pull that inner match out into a helper function, so
that we ban be very explicit about what arguments we take. And now there's
a clear place I can mark with a TODO for the changes needed to unblock
subscript nodes in our AST.

I also move all the helpers out of the `let result = ` block... I have
no idea why we were nesting some of the helpers inside there and others
were outside, it makes no difference and the inconsistent indentation
is kind of confusing.

Reviewed By: samwgoldman

Differential Revision: D57072637

fbshipit-source-id: 78a023a05191f5ee2f045ec84b4b9df1d440b40d