Use a single outer helper for callable type logic
Summary:
It's already a little bit messy that we duplicate some logic across
`is_typing_callable` and `parse_callable` helper functions inside
of `Type.create_logic`.

But once we cut over to subscript expressions instead of using
`__getitem__` call nodes even more things break (because the current
definition of `is_typing_callable` is recursive and actually at the
core agnostic to what kind of call we are making... that cannot be
the case once we cut over).

In preparation for improving the logic, then, it's helpful to extract
both things into a common interface that returns an optional type,
which will be None when the type isn't Callable (in that case we'll
fall back to the usual parametric logic) and `Some <callable_type>`
when it is.

In addition, the `resolve_getitem_callee` logic belongs inside this same
new helper function not floating up in the `Callable` module a thousand lines
earlier:
- it is *only* used here
- it has the same issue of breaking down a callee which won't work
  when we refactor, so it has to be bulk-updated along with the other
  two inner (well, inner as of this commit) helpers.

Reviewed By: samwgoldman

Differential Revision: D57072636

fbshipit-source-id: 00a2f55907e432b1895530ec477522d6d8c0cd85