Don't close over eager module layout in read only API
Summary:
Background:
Pyre parallelizes work by creating a fixed number of worker processes then
sending "jobs" to them over a pipe. The job is a serialized OCaml function which
is executed by the worker. This works because OCaml's `Marshal` module supports
serializing closures.

To serialize a closure, OCaml encodes a function pointer and also any closed-over
values. In practice, Pyre jobs close over the read-only environment. It turns
out that the read-only environment could contain a large data structure: the
eager module tracker layout.

In Pyre projects that use eager module tracking, the layout maps from module
references to file system paths. This mapping was stored as a hash table in the
OCaml process. Because this data was part of the read-only environment, it was
passed to workers for every job.

For large projects, the layout could get quite large, causing unnecessary
overhead when serializing jobs and sending them to workers. This overhead is
significant because sending jobs to workers faster means better worker
utilization and better performance.

This diff:
In this change, I move the lookup table into shared memory and re-write the
`look_up_qualifier` function in the read-only API to rely only on shared memory.
This means we no longer need to pass the mapping to workers. Instead, we only
pass the "handle" of the shared memory table, which is a single `int`.

We need to also keep the mapping in the master process's memory to support
iterating over all module paths, which is not currently possible through shared
memory tables.

This is a large diff (sorry, I can break it up if needed for review), moving a
few of the existing concepts around. The moving pieces are:

1. Extracting a "SharedTable" abstraction, which generalizes the existing shared
memory logic used by the lazy module tracker.
2. When creating the eager module tracker, create and populate shared memory
tables as well as the in-memory table.
3. Change the `find` method in the eager API to rely exclusively on the shared
memory tables.
4. Tweak the definitions used for `look_up_qualifier` to close over only the
find methods for explicit and implicit modules, to avoid accidentally closing
over the in-memory mapping.

Follow-up:
Because this diff changes the overhead costs of sending a job, we might want to
re-evaluate the scheduler policies for the various phases. In particular, we can
investigate sending smaller jobs to workers, which can lead to better worker
utilization by minimizing "long pole" jobs. However, this needs to be balanced
against caching and other factors.

Reviewed By: stroxler

Differential Revision: D57065745

fbshipit-source-id: e2f177a9dbcd2ff3b18fdb6f6a9aa96f2424b46f