Add placeholders for Subscript logic in the easy cases of static analysis
Summary:
I'm now getting close to being ready to add an Expression.Subscript
node to the AST and work on cutting the parser over to produce it
instead of `__getitem__` calls.

It will be easier to do that work if, before making sweeping changes
like that, I first massage the callsites so that the key logic
is already usable, which I can do by just breaking out the
special case of `Call` that is currently produced by the Parser and
ensuring that the only things we ever bind in the match are `base` and
`index`... by doing this I ensure that the body of that match is reusable
for `Subscript` nodes, which makes it much easier to be confident in the
later diff where I have to add all the extra match cases.

This commit does so for the simpler of the Pyre traversals:
- scope.ml
- uninitializedLocalCheck.ml
- unawaitedAwaitableCheck.ml

The other traversals (pysa, type check, global leak check, and preprocessing)
have more significant interactions with `__getitem__` and so I'll handle
them in separate diffs.

Reviewed By: yangdanny97

Differential Revision: D57113191

fbshipit-source-id: e7e48bd1cde67bce8b3136dd04c0a1a00c76905e