Remove unnecessary Fork wrapper module
Summary:
The fork wrapper flushes all buffers before calling fork. This is useful (it's
common to call fflush before fork in C), but the implementation is overkill.

We only call fork in two places: (1) when creating workers and (2) when handling
a job from a worker.

We should flush before (1) creating workers, because the main process likely has
buffered output at this point in the program. For example, the logging output is
buffered. However, we just need to flush once, not before every fork.

We don't need to fork when handing a job, because the worker process will
certainly not have any buffered output, since the worker doesn't do any real
work.

Reviewed By: grievejia

Differential Revision: D57124693

fbshipit-source-id: dafb2990d8aff2433885bd4a9c32a4db70fe08e0