Don't use setsid after forking workers
Summary:
The setsid call was added in D2678741, for a good reason. When you daemonize a
process, you want to disassociate it from the controlling terminal. In this
case, the `flow status` command would sometimes spawn a server process, but we
don't want SIGINT to also terminate the server.

Pyre only uses the Daemon module for workers, which don't have the same
requirements as other daemon-like things. Eventually, I'm just going to merge
the Daemon module into the Worker module to make this much clearer.

Reviewed By: grievejia

Differential Revision: D57146069

fbshipit-source-id: 489e3305dffe3a7bb43baf71219fc16708055c05