Simplify Daemon API
Summary:
The Daemon API includes a type-safe abstraction around the type-unsafe Marshal
module for serializing data across the pipes. While I think this is a genuinely
nice API, it's overkill for what we need.

This diff simplifies the Daemon module to just fork/kill, returning the "raw"
Unix.file_descr values instead of typed channels. I've also updated the Worker
module to create untyped channels and use the type-unsafe Marshal APIs.

My goal here is to remove the Daemon module and just inline the fork/kill
implementations directly in the Worker module, so the type safe API is kind of
useless. In any event, it's not too hard to ensure this code is correct without
the OCaml type system's support.

Reviewed By: grievejia

Differential Revision: D57146060

fbshipit-source-id: 7b77a1a68eac375ff0c8aa22a264e02cd69c98da