Simplify the union type handling in preprocessing.ml
Summary:
I was working on a diff to prepare the code in `preprocessing`
so that it would be easier to cut over to subscripts; there are
a lot of mostly small changes needed that I've grouped together
in a diff.

But the union logic is extremely messy, and deserves its own stand
alone cleanup; we destructure unions multiple times and we were
also defining a helper used in one place dozens of lines away from
it.

All of this makes it much harder to clean up the logic, so to make
this commit I first inlined the `union_value` function and then
worked to simplify all of the surrounding plumbing.

Reviewed By: grievejia

Differential Revision: D57171312

fbshipit-source-id: af5193de297678588c74f3bd7461ca7cb8eb677a