Do not import tests in other tests!
Summary:
Doing this breaks isolation of tests at the level of *file* isolation, which
is a disaster for iteration because it's impossible to even list tests without
running them, much less isolate them.

It's much better to just duplicate a little helper than to import it and wind
up running the test suite multiple times.

Reviewed By: grievejia

Differential Revision: D57171314

fbshipit-source-id: cc58442199c3fff5115c938918ec4951a52fd78d