Prepare preprocessing.ml for subscripts
Summary:
In order to make it reasonably feasible to introduce a `Subscript`
AST node, I want to first make as much downstream logic as possible
compatible with `Subscript`, meaning we need to
- always match on `__getitem__` separately from the "general" call case
- the body of the match should only make use of `base` and `index`, not
  any other aspec to fthe call (because that's all we'll get from
  `Subscript`)

With that done in advance, cutting over will be a simple mechanical
transformation of the code so that even though we may need large diffs
they should be simple.

This commit prepares `presprocessing.ml` for
this: I break out the `__getitem__` case everywhere using only `base`
and `index`. In one place where there's nothing to do, I left a
`TODO` marker with `__getitem__` in the comment because it's important
that I can easily find that and support `Subscript` when searching for
`__getitem__`.

Reviewed By: grievejia

Differential Revision: D57113192

fbshipit-source-id: 58136d95407a546451018efbd6dec9c7f32304c6