Add upgrade tests to test script (#847)
Summary:

Internal tests will never catch the use of Python features
(like the pipe operator on types without `__future__.annotations`)
that are incompatible with older versions of Python that we still
need to support.

As a result it's very important that we cover this in gituhub tests;
even though we frequently break CI, this at least makes it easy to
prep for an open-source release: first check CI for test failures
on any supported Python version and fix issues, only then start a
release.

Unfortunately the script wasn't running upgrade tests, and as a result
didn't catch an invalid use of the pipe operator in unguarded type
annotations.

Differential Revision: D57218140