Split test cases (where it's easy) in typeAliasEnvironmentTest
Summary:
This diff is motivated by the need to debug alias environment issues that
popped up during flipping Pyre to use Subscript nodes, but it's a good
BE fix in general - easier test isolation means much better iteration
cycles.

The update tests are serial and share a project, so those remain one long
chain of assertions and likely need to stay this way.

Reviewed By: yangdanny97

Differential Revision: D57189746

fbshipit-source-id: 524255fb450291f44f9781ffb77724c2794cc905