Split a few test cases in `typeTest.ml`
Summary:
I usually like to split test cases in an entire module, but
`typeTest.ml` is huge and this would take me a long time
(probably on the order of an hour) which isn't a good investment
right now.

So I'm splitting just the tests that I need to for debugging
subscript changes to the parser; I'll split others whenever
I run into a need.

Reviewed By: grievejia

Differential Revision: D57175699

fbshipit-source-id: 2529c1681f0b50de00515cb0d7b66bec28514aa2