Prepare `modelParser.ml` for subscripts
Summary:
In order to make the actual cutover from `__getitem__` calls to
subscripts - whcih consists of a handful of enormous diffs - as simple
as possible, I'm trying to prepare ahead of time by making sure that
all of the uses destructure the `__getitem__` call in a way that is
compatible with `Subscript`: we should always handle it separately
from "general" call cases and we should only bind `base` and
`index`, where `index` is the expression of the one positional
argument.

This diff is making that change in `modelParser.ml`; most of the
changes are minimal but the handling of `parse_sanitize_annotations`
had the wrong type signature; in order to handle both arguments
coming directly out of a decorator (which will continue to be
a list of `Call.Argument.t`) and also an index from a subscript
(which will no longer be a call argument) we need to make it operate
on expressions rather than arguments.

Reviewed By: tianhan0

Differential Revision: D57171310

fbshipit-source-id: ff959c30a9e6039ab0ddea4fb9e478a4c56786b5