Prepare globalLeakCheck for Subscript AST nodes
Summary:
This commit re-wires the globalLeakCheck.ml logic to be
forward-compatible with Subscript nodes where all we will
have is a `base` and `index`, no call.

I change the behavior slightly here: I no longer consider
globals to be accessible via indexes. The test case I changed
had a TODO marker on it saying we should modify this behavior
anyway, so I think this is a net win :)

Reviewed By: grievejia

Differential Revision: D57171311

fbshipit-source-id: 42cf0d5abc4568252c5cf6253b07439e1adf2102