Create a `redirect_expressions` hook in Callgraph.ml
Summary:
Pysa currently has some hooks to rewrite the source coming from Pyre:
- it has long used rewrites of special calls via Pyre's
  `redirect_special_calls`
- recently in D56883262 we added the ability to redirect entire
  statements, so that we could preserve in Pysa the legacy behavior
  of rewriting subscript assignments into `__setitem__` while eliminating
  that behavior (and the bugs on multi-target assignment that followed)
  from Pyre.

Because of the way Pysa relies on callgraph registration, we need similar
support going forward for special calls that are created not from a call
but rather from a different kind of Expression node: in particular, I will
shortly be working on eliminating `__getitem__` lowering from the parser.

So, we need this rewrite to be more general - it should be able to
transform expressions rather than only working on Call nodes.

This commit makes that change in the signature, without actually modifying
the behavior (for now, the implementation of `redirect_expressions` is just
to call `redirect_special_calls` on calls).

Reviewed By: tianhan0

Differential Revision: D57113189

fbshipit-source-id: b8a2f0fffe9b230b9f79024351bde06384ab3e41