Make redirect_assignments forward-compatible with Subscript
Summary:
The diff adding a Subscript node and cutting the most essential use cases
over to supporting it will be huge, so I want the transformation to be
as simple / mechanical as possible.

As a result, I want the existing matches on `__getitem__` calls to always
use only `base` and `index` in the match, we don't want to be assuming the
index comes in as a `Call.Argument.t` or anything else that won't translate
well.

By doing this, we'll be able to reuse the same body for the Subscript match
statements in the first part of the cutover, making it pretty obvious that
the cut-over logic is correct which is really helpful given that it will
be a huge, sprawling commit.

Reviewed By: tianhan0

Differential Revision: D57171315

fbshipit-source-id: a4649ff85ee1b04d933cf775a91629162def3bc1