Don't use patch for pyre in upgrade tests
Summary:
Our uses of `patch(...)` are broken for open-source, because
in our current open-source CI our `__name__` involves `pyre-check`,
which is not legal to use.

As a result, for patching anything internal we need to use
`patch.object` instead.

It's okay to use `patch(...)` for stdlib, that doesn't cause
problems (although I do wish we didn't have to rely on patching
quite so much!)

I may break this into a couple commits to keep them small.

Differential Revision: D57358329