Disable a test that is broken on github
Summary:
It's a little unfortunate to disable a unit test that passes
when run using buck, but we really can't have tests checked
in that break github CI.

We need to figure out why this breaks, and either fix it or
find a way to run it only locally, I filed T189132787 to
look at this. In the meantime though I think the best option
is to disable the test because folks are getting spammed with
github CI failure notices.

Differential Revision: D57361655