Refactor worker entrypoint to prepare for long-lived workers
Summary:
Context:
Long-lived workers will change Pyre's parallelism model. Currently, Pyre creates
worker processes which listen for jobs on a pipe. For each job, the worker
forks, performs the job in that forked process which then exits.

We want to prepare for OCaml 5 where we can run parallel work in domains
(threads) in a single process. To prepare for this, we first migrate from the
forking worker model to a long-lived worker model.

Under long-lived workers, we will be able to find and fix any memory leaks and
tweak caching for memory use.

This diff:
This refactor makes it possible to call `worker_loop` with different handlers.
You can either use `fork_handler` to run jobs in a fork or use `worker_job_main`
to run the job in the worker directly.

Reviewed By: stroxler

Differential Revision: D57280730

fbshipit-source-id: 7e94628cafc17d9a78fd3d1b81d7462ce95c5f4a