Remove WorkerCancel
Summary:
Context:
Long-lived workers will change Pyre's parallelism model. Currently, Pyre creates
worker processes which listen for jobs on a pipe. For each job, the worker
forks, performs the job in that forked process which then exits.

We want to prepare for OCaml 5 where we can run parallel work in domains
(threads) in a single process. To prepare for this, we first migrate from the
forking worker model to a long-lived worker model.

Under long-lived workers, we will be able to find and fix any memory leaks and
tweak caching for memory use.

This diff:
Pyre does not use the WorkerCancel module at all. Removing support for this
makes it clear that we do not need to support cancelation from a long-lived
worker.

For the curious, canceling workers happens when the main process sets a flag in
shared memory. Workers need to periodically check this flag, and will currently
exit when they observe the set flag.

Supporting worker cancelation for long-lived workers is certainly possible.
Instead of exiting, workers could raise an exception on cancelation which is
then handled by the top level.

However, since Pyre does not make use of worker cancelation, it's easier to just
remove this code and be sure that we will not exit from the long-lived worker.
Eventually, we should add worker cancelation back in, but we can do that after
migrating to OCaml 5.

Reviewed By: grievejia

Differential Revision: D57178493

fbshipit-source-id: 666a7319a97753600fe65fc7a0d1079d37c24ff9