Add configuration to enable long-lived workers
Summary:
Context:
Long-lived workers will change Pyre's parallelism model. Currently, Pyre creates
worker processes which listen for jobs on a pipe. For each job, the worker
forks, performs the job in that forked process which then exits.

We want to prepare for OCaml 5 where we can run parallel work in domains
(threads) in a single process. To prepare for this, we first migrate from the
forking worker model to a long-lived worker model.

Under long-lived workers, we will be able to find and fix any memory leaks and
tweak caching for memory use.

This diff:
We can now switch between forking and long-lived workers with this flag. Next
diff I will connect this to the Pyre client binary.

Reviewed By: stroxler

Differential Revision: D57280731

fbshipit-source-id: 546f813329a5113101103f52692e2215af8a74ab