Make `Expression.dequalify` use an exhaustive match
Summary:
We often avoid exhaustive matching in Pyre for things like Expression,
which I think is a reasonable tradeoff much of the time when a function matches
one or two cases; it would be hard for readability to capture all the others
(I actually think having text editors understand fallback and auto-hide the
other cases might be better, but that tooling doesn't exist as far as I know).

But in the case of `delocalize`, it can be *really* hard to find a breakage caused
by failing to update this so I think it's actually worth the cost of doing a full
match. I spent several hours debugging issues when switching from `__getitem__`
to `Subscript` nodes where the root cause was a missed update to this match.

We probably should move this function to `analysis` and also in general centralize
more of the qualification and de-qualification logic so that it's easier to see all the
interactions in one place, but that's not really in scope for my current subscript
lowering push and I don't want to get sidetracked.

I honestly don't know exactly why `delocalize` operates on only a few cases, I *think*
that it's maybe supposed to work only on annotation expressions and not more general
expressions but there are no docs which means that's just a guess.

Reviewed By: grievejia

Differential Revision: D57295180

fbshipit-source-id: 9ed519ce486d6ea6c829a5d940ca8260913025c5