The big Subscript cutover [1/n] new AST node
Summary:
This diff starts the big cutover from the parser producing
`__getitem__` calls on subscript expressions like `base[index]` to
producing `Subscript` nodes instead.

The first diff of the cutover that can't be broken into small
work is adding the Subscript node to the AST itself, which requires
updating all of the plumbing that traverses ASTs (transformers,
folders, visitors) as well as every place in the analysis that does
a full pattern match on `Expression.expression`.

I'm still trying to chunk the work, which means I'm putting off the
following in this diff and will stack the work:
- supporting `Subscript` in 100% of the `__getitem__` uses: this diff is really
  just cutting over all of the modules that *had* to be cut over because they
  do a full match somewhere. There are quite a few other places we need to cut
  over (e.g. Type.ml's expression-to-type converter), and I'll stack as many of
  them as I can easily find in the next diff; there's probably going to be a
  long tail I don't find until I cut over the CPython parser and get test
  failures.
- cutting over the parsers themselves, as well as the Preprocessing.ml
  code that produces synthetic subscript expressions.
- eliminating all the legacy `__getitem__` matches... that has to wait
  until the very end after all parsers + preprocessing are cut over.

Reviewed By: yangdanny97

Differential Revision: D57343980

fbshipit-source-id: 82f8c9a92689385c128b3878ae022e14f99567f7