Make protocol handling logic slightly less brittle
Summary:
The logic in `classHierarchyEnvironment` for extracting base type
generic information appears to depend heavily on `Type.expression`,
which seems... very confusing.

A (maybe accidental?) side-effect of this is that we do not consider classes
that inherit from `typing.Protocol` to be generic:
- in `compute_generic_base` -> `extract_protocol_parameters`, we detect
  when there's a Protocol and we force the parameters to be a list instead
  of none
- then, in `Type.expression` we *generate an illegal __setitem__ call*
  if the parameters are `None`
- then, when we parse our illegally-generated expression for our
  illegally-constructed generic type, we get `Type.top` because parsing
  fails, and we consider it to be non-generic.

This flow of logic is completely bonkers.

Why not just match on the type and if it's a *generic* protocol then we
get its parameters, otherwise we don't? I confirmed that without this change,
`globalResolutionTest` fails because we actually can parse the generated
expression and we get empty parameters, incorrectly marking the class as
generic. With this change, the regression no longer happens.

NOTE: we're still using the seemingly-crazy convert-expression-to-type-and-back
logic here because I'm not yet sure why that exists, and I'm not ready to eliminate
it just yet. But at least we're no longer relying on this generating a type parsing
error! I filed T189106851 to look into this more later.

Reviewed By: yangdanny97

Differential Revision: D57348832

fbshipit-source-id: 0872b8362eab958edc1bb7a951a8b4fbb9b35666