Split cases in annotatedSignatureTest
Summary:
This looks like one of the easiest places for me to start
digging on one fo the bugs still popping up in the change to
have Subscript AST nodes.

To debug effectively, I need to have the test cases split
so that it's easy to read failures and isolate tests.

Reviewed By: yangdanny97

Differential Revision: D57382723

fbshipit-source-id: 27e338a764b1c484c82e6f1328a254431c362b84