Update full_result.json
Summary:
This updates to what we are currently seeing in the integration tests.

There is a regression here: we don't seem to be resolving
werkzeug.utils.redirect properly; Maxime was able to verify that
reveal_type(redirect) comes back as `unknown`.

It's likely that this is either Pyre not finding the package due to some
configuration handling change, or a signature extraction issue.

I filed T189346515 to investigate, but it seems wise to make CI pass
in the meantime so that we'll notice the next regression quicker.

Differential Revision: D57437213