Handle package changes in github CI
Summary:
The dataclasses-json version we are getting in github has types that
are incompatible with the internal version. As a result, we need to pin
the version on github to get comparable results (in addition it's at
least possible we would get bugs in open-source without pinning). I
filed T189226004 to bump the version internally.

In addition:
- we aren't marking `toml` as a dependency, but we need it as a dev
  dependency because type checking the typeshed patcher requires it
- we need to list both tabulate and toml in our .pyre_configuration
  for github to pick them up because we are still using the legacy
	setup where packages aren't picked up by default... this would
	probably be good to change at some point.

Reviewed By: arthaud

Differential Revision: D57387918

fbshipit-source-id: a45d0236dbcf28788c08bedbf87dd5fd696c0faf