Remove unnecessary heap handle, simplify hh_connect
Summary:
Now that we fork workers instead of spawning, it's not necessary for workers to
"connect" in the same way. All memory mappings from the main process already
exist in the worker so we don't need to mmap again or initialize globals.

Instead, we just need to set the `my_pid` global. We can also simplify the
initialization code by inlining helper functions that are now only called in one
place.

Reviewed By: grievejia

Differential Revision: D57398902

fbshipit-source-id: a5f4a962d11134fd803292e0c713637163048ff6