Use anonymous shared memory
Summary:
Now that Pyre only uses fork for workers, we can use an anonymous memory mapping
for shared memory. This greatly simplifies this code, since we no longer need to
create a file descriptor to pass through the exec to another process.

The code to create a fd for shared mem was particularly complicated because APIs
for this are not very portable. We used memfd_create on Linux (where possible,
it's a somewhat recent API) or SysV shm_open, with a fallback to manually
configured directory (hopefully memory-backed like tmpfs).

Instead, we can use the much more widely available MAP_ANON flag to mmap,
combined with MAP_SHARED so the child process can see each other's writes.

Reviewed By: grievejia

Differential Revision: D57416765

fbshipit-source-id: 92bf791fdebf90447b3545cc68d6ca136980b24c