Improve AnnotatedSignatureTest readability
Summary:
Remove a lot of nesting nesting, give some variables clearer names, and move some
of the computation to where it is used.

All of this is because I needed to be able to debug the guts of a failing
test, and this test consists of some *highly* nontrivial jumps where we
synthesize a fake source, extract parsed statements and expressions from
the source, and then do things with them in isolation.

Flattening and renaming makes it much easier to see what's actually getting
passed around, and clearer names help (especially since we're literally
passing a *fragment* of a callable type expression / a *fragment* of a
call expression as the args, so using nondescriptive variable names
like `callable` and `arguments`, that then get shadowed repeatedly in
the test body, is pretty confusing).

Reviewed By: grievejia

Differential Revision: D57489050

fbshipit-source-id: 5446511c202b25e2442df025d70f15d16301589f