Bugfix: Make synthetic getitem `special = true`
Summary:
I'm breaking this out in a separate diff in part so that it's
clear in the history that this is very important: it turns out AttributeResolution
will create an empty attribute table if both `accessed_through_class` and `special_method`
are set to `True`, which will trigger us to fall all the way back to the metaclass
implementation of an attribute.

This works in practice, but it's an incredibly roundabout way of going and it
was completely undocumented; everywhere else in Pyre, `special = true` is a
marker used to indicate when an *AST Node* was made up by the parser from
built-in syntax (for example, binary operators get `special = true` and they
have absolutely nothing to do with this code path).

The end effect of *not* having this in place is that if you try to do
`resolve_expression` on something like `list[str]`, Pyre will wind up
resolving it to a "normal" `list.__getitem__`, whereas we actually need
it to resolve to `GenericMeta.__getitem__` in order for the magical
`instantiate_attribute` logic in `AttributeResolution` to properly intercept
the call.

What's particularly inanane about this is that `GenericMeta` doesn't even exist
anymore - if you look at `typing.pyi` it isn't there:
https://github.com/python/typeshed/blob/main/stdlib/typing.pyi

We are actually forced to "invent" this class (which disappeared in 3.7!!)
in `missingFromStubs`, because Pyre's logic to handle generic subscripts
depends on this thing that hasn't existed for years.

OK, end of my half-rant half-explanation... we definitely should look
into this more though, because the whole class of bug where Pyre complains
about `GenericMeta.__getitem__` when people use perfectly valid type expressions
is tied in with this.

Reviewed By: grievejia

Differential Revision: D57489049

fbshipit-source-id: af8e70757ea2f6fad1726fb4d2e14b28cb8a438c