The big Subscript cutover [2/n] handle Subscripts in most places
Summary:
The previous diff added a new `Subscript` node to Pyre and handled it in
all of the places where the compiler made it really easy to see we had to,
especially the static analyses that match on `Expression` as a whole.

This commit follows up on that by adding hooks to handle `Subscript` everywhere
else that I know of where we need it, mainly all the places like `Type.create_logic`
and Pysa model parsing where we've baked `__getitem__` into our expression-to-type
code.

This is a big diff, but the goal is to have a forward/backward compatible state that lets me
cut over the parsers and preprocessing code in as small a diff as possible,
with the ability to see before-and-after code at the same time while I'm debugging
issues that pop up

NOTE: I was trying to get all the places we need subscript in this diff, but to unblock
binary operator AST nodes we may need to land this before I've figured everything
out, in which case a few diffs may have to be stacked on top.

Reviewed By: grievejia

Differential Revision: D57343978

fbshipit-source-id: 5e81685b15f8a9cc6cda3a1a6ad18bbf705c4c8d