Bugfix: add Subscript to `get_identifier_base`
Summary:
This is another super gnarly bugfix for subscript logic based on surprising
AttributeResolution behavior: it turns out that if AttributeResolution sees a
global that is a type alias, it will will not actually use `TypeAlias` as the
type but rather it will attempt to infer the type from scratch.

There are some reasons for this - for example it is not legal to call a type
alias as a constructor, but it works at runtime and the spec is ambiguous about
whether type checkers should support it; our current logic will indeed allow
it.

But this means there's a very special hack inside of `resolve_literal` in an
inner function called `fully_specified_type` where we special-case subscripts
whose base is a simple name (these are the ones that could plausibly represent
generic types).

I updated the inner match correctly, but I missed two blocks of filtering logic:
- we only call `fully_specified_type` when `has_identifier_base` is True, which relies
  on `Expression.get_identifier_base`, which I forgot to update `/facepalm`
- we also have a second layer of match on the same expression further down,
  and we only call `fully_specified_type` against `Call` nodes; we actually
  should only be doing this on `Subscript` nodes.

This diff fixes both issues, causing Pyre to continue the existing behavior
of treating implicit type aliases in a sort of "dual" way where they can behave like
aliases or like concrete type object values.

This may not actually be a desirable behavior, but backing away from it requires
care for two reasons:
- we need to understand how many exising uses rely on this
- we also need to understand how trustworthy TypeAliasEnvironment really is;
  it will treat any expression that "might plausibly be a type" based on pure
  syntax as "definitely a type", which means in general it could be pretty
  wrong to infer TypeAlias whenever TypeAliasEnvironment is willing to.

Reviewed By: grievejia

Differential Revision: D57489052

fbshipit-source-id: 84e333636464b1fd9b16c5ce70d34548b42464e6