Handle Subscript in implicit __getitem__ match for Sanitize
Summary:
My earlier work to cut over all of the findable uses of `__setitem__` failed to
catch this case, where we are actually matching a stringified Expression and then
assuming it must be a `Call`.

I had to step through for a while to find it, but the fix is easy - support
Subscript here, and later in a clean up diff we can remove the `Call` match.

Reviewed By: grievejia

Differential Revision: D57489048

fbshipit-source-id: f5a6c2c6b54e9be043adfd5cd3e2bd53c9771c5a