Don't error when trying to kill workers that have already exited
Summary:
`Scheduler.with_scheduler` will try to kill all workers on any exception. If a
worker exits unexpected, that results in an exception, which means we try to
kill a worker that is already dead, which causes an exception to be thrown from
the finally handler of `with_scheduler`, clobbering the original stack trace.

This diff wraps operations that might fail with exception handlers, so the
original exception propagates, giving better information about what caused the
crash.

This change was useful when investigating a specific case where workers crashed,
but I'm not convinced this is the best solution. Alternatively, we could
eliminate the `with_scheduler` API and let exceptions thrown propagate and kill
the main process. This would ensure that any child processes are cleaned up as
well. I'm mentioning this alternative, but have no plans to implement that
change.

Reviewed By: stroxler

Differential Revision: D57363940

fbshipit-source-id: d0529cc9605b3088e4d0af056416fff459c5d25f