Flush statistics state before forking worker processes
Summary:
If the statistics module's cache has data in it, we need to flush that data
before creating workers. Otherwise, the workers will all redundantly flush the
same data.

This bug was introduced in D56831118, where we changed to forking workers
instead of spawning them.

The need to flush things before forking was known (see D57124693), and we do
properly flush IO buffers, but I was not aware of this state which also needed
to be flushed.

Reviewed By: stroxler, yangdanny97

Differential Revision: D57422520

fbshipit-source-id: 31c00f4ebac58a45d493db0dd0a1465fc307e73e