The big Subscript cutover [3/n] Cut over parsers + preprocessing.
Summary:
This commit is the main cutover for getting rid of `__getitem__`
lowering from the parser.

Up until this point, we
- did all of the work (a lot of work!) refactoring the core logic
  itself to be compatible with the future AST
- added a `Subscript` node and extended the analysis code to handle
  both it and the old `Call` nodes

In this commit, we flip all of the actual code generating AST nodes
from generating the `Call { callee = { ... "__getitem__" ...}}` nodes
to `Subscript { base; index }` nodes.

By "actual code generating AST nodes" I mean:
- all three parsers, as well as
- the preprocessing code wherever we create a synthetic value
- other synthetic value creation, for example in `type.ml` type-to-expression code

... these have to be cut over together because we validate Preprocessing.ml
against the output of the Menhir parser, so doing it in any smaller chunks
that a single cutover diff would mean we have to go against the grain to keep
`make test` working

Reviewed By: grievejia

Differential Revision: D57343977

fbshipit-source-id: 620596c03070e3f375727c68c7355662c0df1f76